Python/requirements.txt
Flask==2.3.3
Werkzeug==3.0.1
flask-cors==4.0.0
gspread==6.0.2
oauth2client==4.1.3
google-cloud-secret-manager==2.18.1
firebase-admin==6.4.0